% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HSSM_functions.R
\name{BO_MomEst}
\alias{BO_MomEst}
\title{Moment estimator - BO}
\usage{
BO_MomEst(
  n_j,
  K12,
  S12,
  Kmin = 1,
  Kmax = 100,
  Smin = 0,
  Smax = 100,
  gamma_LB,
  gamma_UB,
  lambda_LB,
  lambda_UB,
  BO_iterations = 50,
  Max_iter = 100
)
}
\description{
Moment estimator - BO
}
