% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HSSM_functions.R
\name{BO_MomEst_multiple_NB}
\alias{BO_MomEst_multiple_NB}
\title{Moment estimator - BO}
\usage{
BO_MomEst_multiple_NB(
  n_j_list,
  v_K12,
  v_S12,
  Kmin = 1,
  Kmax = 100,
  Smin = 0,
  Smax = 100,
  gamma_LB,
  gamma_UB,
  mu0_LB,
  mu0_UB,
  r_LB,
  r_UB,
  BO_iterations = 100,
  Max_iter = 100,
  p = 2,
  c1 = 0.5,
  c2 = 0.5
)
}
\description{
Moment estimator - BO
}
