% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HSSM_functions.R
\name{BO_MomEst_NB}
\alias{BO_MomEst_NB}
\title{Moment estimator - BO}
\usage{
BO_MomEst_NB(
  n_j,
  K12,
  S12,
  K1,
  K2,
  Kmin = 1,
  Kmax = 100,
  Smin = 0,
  Smax = 100,
  gamma_LB,
  gamma_UB,
  mu0_LB,
  mu0_UB,
  r_LB,
  r_UB,
  BO_iterations = 50,
  Max_iter = 100,
  perr = 2,
  pesi = c(0.25, 0.25, 0.25, 0.25)
)
}
\description{
Moment estimator - BO
}
