% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compute_log_Pochhammer}
\alias{compute_log_Pochhammer}
\title{Pochhammer log Symbol}
\usage{
compute_log_Pochhammer(x, a)
}
\description{
\loadmathjax This function computes the Pochhammer symbol in log form. See \code{\link{compute_Pochhammer}} for details.
}
