% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compute_log_Vpost}
\alias{compute_log_Vpost}
\title{log V post}
\usage{
compute_log_Vpost(r, k, m_i, n_i, gamma, prior, prior_param, M_max)
}
\arguments{
\item{r:}{number of NEW species}

\item{k:}{number of OLD species}
}
\description{
log V post
}
