% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compute_log_Vprior}
\alias{compute_log_Vprior}
\title{log V prior}
\usage{
compute_log_Vprior(r, n_i, gamma, prior, prior_param, M_max)
}
\arguments{
\item{r:}{number of global distinct species}
}
\description{
log V prior
}
