% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compute_logC}
\alias{compute_logC}
\title{compute_logC - Compute log of absolute values of non Central C number}
\usage{
compute_logC(n, scale, location)
}
\arguments{
\item{scale}{must be strictly negative.}

\item{locatio}{must be non positive. Set to 0 for central C numbers.}
}
\description{
\loadmathjax This is the main function in the computation of C numbers. It uses the (2.69) formula in the "Combinatorial methods in discrete distributions" book.
It computes \mjseqn{log(|C(n,k; scale, location)|)} for each k=0,...,n.
This implementation uses Rcpp vectors.
}
