% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compute_Pochhammer}
\alias{compute_Pochhammer}
\title{Pochhammer Symbol}
\usage{
compute_Pochhammer(x, a)
}
\description{
\loadmathjax This function computes the Pochhammer symbol,
\mjsdeqn{(a)^{x} = \frac{\Gamma(a+x)}{\Gamma(a)}}.
Where \code{x} is a real number. When x is an integer, such a function coincides with the rising factorial defined in \code{\link{raising_factorial}}.
The raising (here denote with the upper apex) and the falling factorial (here denote with the lower apex) are related by the following relationship
\mjsdeqn{(a)_{n} = (-1)^{n}(a)^{n}}.
}
