% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HSSM_functions.R
\name{Corr_Diversity}
\alias{Corr_Diversity}
\title{Correlation and diversity}
\usage{
Corr_Diversity(gamma_j, lambda, B = 10000)
}
\description{
Compute via Monte Carlo estimation the correlation and the inverse of the Simpson diversity
}
