% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HSSM_functions.R
\name{Corr_GiniSimpson}
\alias{Corr_GiniSimpson}
\title{Correlation and Gini-Simpson index}
\usage{
Corr_GiniSimpson(gamma_j, lambda, B = 10000)
}
\description{
Compute via Monte Carlo estimation the correlation and Gini-Simpson index
}
