% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HSSM_functions.R
\name{D_distinct_post_interval}
\alias{D_distinct_post_interval}
\title{Compute posterior distribution}
\usage{
D_distinct_post_interval(
  m_j,
  n_j,
  k,
  gamma,
  prior = "Poisson",
  ...,
  Max_iter = 100,
  Kmin,
  Kmax,
  logVpost_vec = NULL,
  print = TRUE
)
}
\description{
Compute posterior distribution
}
