% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{D_distinct_post_interval_c}
\alias{D_distinct_post_interval_c}
\title{Compute the distinct posterior distribution in a given interval. Hence, in general, the returned values may not sum up to 1}
\usage{
D_distinct_post_interval_c(
  m_j,
  gamma_j,
  n_j,
  k,
  prior,
  prior_param,
  M_max,
  Kmin,
  Kmax,
  logVpost_vec,
  print
)
}
\description{
Compute the distinct posterior distribution in a given interval. Hence, in general, the returned values may not sum up to 1
}
