% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HSSM_functions.R
\name{D_distinct_prior}
\alias{D_distinct_prior}
\title{Compute prior distribution}
\usage{
D_distinct_prior(
  n_j,
  gamma,
  prior = "Poisson",
  ...,
  Max_iter = 100,
  Kexact = 100
)
}
\description{
Compute prior distribution
}
