% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HSSM_functions.R
\name{D_distinct_prior_interval}
\alias{D_distinct_prior_interval}
\title{Compute prior distribution}
\usage{
D_distinct_prior_interval(
  n_j,
  gamma,
  prior = "Poisson",
  ...,
  Max_iter = 100,
  Kmin,
  Kmax,
  logV_vec = NULL,
  print = TRUE
)
}
\description{
Compute prior distribution
}
