% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{D_distinct_prior_interval_c}
\alias{D_distinct_prior_interval_c}
\title{Compute the distribution for the prior number of distinct components in a given interval. Hence, in general, the returned values may not sum up to 1}
\usage{
D_distinct_prior_interval_c(
  n_j,
  gamma_j,
  prior,
  prior_param,
  M_max,
  Kmin,
  Kmax,
  logV_vec,
  print
)
}
\description{
Compute the distribution for the prior number of distinct components in a given interval. Hence, in general, the returned values may not sum up to 1
}
