% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HSSM_functions.R
\name{D_joint_post_largen}
\alias{D_joint_post_largen}
\title{Compute posterior distribution large n}
\usage{
D_joint_post_largen(
  m_j,
  n_j,
  k,
  gamma,
  prior = "Poisson",
  ...,
  Max_iter = 100,
  Kmin,
  Kmax,
  Smin,
  Smax,
  log_Vprior,
  print = TRUE
)
}
\description{
Compute posterior distribution large n
}
