% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HSSM_functions.R
\name{D_joint_post_square}
\alias{D_joint_post_square}
\title{Compute posterior distribution}
\usage{
D_joint_post_square(
  m_j,
  n_j,
  k,
  gamma,
  prior = "Poisson",
  ...,
  Max_iter = 100,
  Kmin,
  Kmax,
  Smin,
  Smax,
  logVpost_vec = NULL,
  print = TRUE
)
}
\description{
Compute posterior distribution
}
