% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{D_joint_post_square_c}
\alias{D_joint_post_square_c}
\title{Compute the joint posterior distribution in a given square. Hence, in general, the returned values may not sum up to 1}
\usage{
D_joint_post_square_c(
  m_j,
  gamma_j,
  n_j,
  k,
  prior,
  prior_param,
  M_max,
  Kmin,
  Kmax,
  Smin,
  Smax,
  approxC,
  logVpost_vec,
  print
)
}
\description{
Compute the joint posterior distribution in a given square. Hence, in general, the returned values may not sum up to 1
}
