% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HSSM_functions.R
\name{D_joint_prior}
\alias{D_joint_prior}
\title{Compute prior distribution}
\usage{
D_joint_prior(
  n_j,
  gamma,
  prior = "Poisson",
  ...,
  Max_iter = 100,
  Kstart = 1,
  logV_vec = NULL
)
}
\description{
Compute prior distribution
}
