% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HSSM_functions.R
\name{D_joint_prior_square}
\alias{D_joint_prior_square}
\title{Compute prior distribution}
\usage{
D_joint_prior_square(
  n_j,
  gamma,
  prior = "Poisson",
  ...,
  Max_iter = 100,
  Kmin,
  Kmax,
  Smin,
  Smax,
  logV_vec = NULL,
  print = TRUE
)
}
\description{
Compute prior distribution
}
