% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{D_jointKS_post_largen_c}
\alias{D_jointKS_post_largen_c}
\title{Compute the distinct posterior distribution in a given interval. Hence, in general, the returned values may not sum up to 1}
\usage{
D_jointKS_post_largen_c(
  k,
  m_j,
  n_j,
  gamma_j,
  prior,
  prior_param,
  Kmin,
  Kmax,
  Smin,
  Smax,
  log_V,
  M_max
)
}
\description{
Compute the distinct posterior distribution in a given interval. Hence, in general, the returned values may not sum up to 1
}
