% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HSSM_functions.R
\name{Distinct_Prior_MCMC}
\alias{Distinct_Prior_MCMC}
\title{Distinct_Prior_MCMC}
\usage{
Distinct_Prior_MCMC(
  Niter,
  n_j,
  gamma_j,
  part_init = rep(0, sum(n_j)),
  prior,
  prior_param,
  M_max = 1000,
  seed = 0
)
}
\description{
Distinct_Prior_MCMC
}
