% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HSSM_functions.R
\name{Expected_posterior}
\alias{Expected_posterior}
\title{Expected_posterior}
\usage{
Expected_posterior(
  k,
  m_j,
  n_j,
  gamma,
  type,
  prior = "Poisson",
  ...,
  Max_iter = 100,
  tol = 10^-12
)
}
\arguments{
\item{k}{integer, the number of distinct species that have been observed in the sample of sizes given by \code{n_j}.}

\item{m_j}{an positive integer in the case of exchangeable data or a vector of size \code{d} in the case of partially exchangeable data.}

\item{n_j}{an positive integer in the case of exchangeable data or a vector of size \code{d} in the case of partially exchangeable data.}

\item{gamma}{real valued, it must be of the same size of \code{n_j}}

\item{type}{string, select \code{distinct} to compute the expected value of the number of distinct species or select \code{shared} to compute the expected value of the number of shared species.}

\item{prior}{a string that indicates the type of prior to be used for the number of components. It can only be equal to \code{"Poisson"} or \code{"NegativeBinomial"}.}

\item{...}{the addition parameters to be used in the prior. Use \code{lambda} for the "Poisson"case (must be strictly positive) and \code{r} (positive integer) and \code{p} (real in (0,1)) for the "NegativeBinomial" case.}
}
\description{
This function computes the expected value and the variance of the a posteriori probability of distinct or shared species.
}
