% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HSSM_functions.R
\name{Freq_probSh_1step}
\alias{Freq_probSh_1step}
\title{Frequentist estimators}
\usage{
Freq_probSh_1step(data)
}
\arguments{
\item{data}{M x 2 matrix with rownames indicating the species names. Unobserved species must also be included}
}
\description{
Return the frequentist estimators for the 1step ahead probability of observing new shared species
}
