% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HSSM_functions.R
\name{GiniSimpson_est}
\alias{GiniSimpson_est}
\title{Empirical Gini-Simpson}
\usage{
GiniSimpson_est(data_j)
}
\description{
Compute the empirical estimate of the Gini-Simpson index via iNEXT package function
}
