% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{log_eppf}
\alias{log_eppf}
\title{Compute the log of the eppf under Poisson prior. This works only in the exchangeable case (d=1)}
\usage{
log_eppf(n, r, n_k, gamma, Lambda, M_max)
}
\description{
Compute the log of the eppf under Poisson prior. This works only in the exchangeable case (d=1)
}
