% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{log_peppf}
\alias{log_peppf}
\title{Compute the log of the peppf under Poisson prior}
\usage{
log_peppf(r, n_j, n_jk, gamma1, gamma2, Lambda, M_max)
}
\description{
Compute the log of the peppf under Poisson prior
}
