% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{log_raising_factorial}
\alias{log_raising_factorial}
\title{log Raising Factorial}
\usage{
log_raising_factorial(n, a)
}
\description{
\loadmathjax This function computes the logarithm of the rising factorial \mjseqn{(a)^{n}} using the gsl code for the log of Pochhammer symbol.
See \code{\link{raising_factorial}} and \code{\link{compute_Pochhammer}} for details.
}
