% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{log_Vpost_long}
\alias{log_Vpost_long}
\title{All terms of V}
\usage{
log_Vpost_long(r, k, m_i, n_i, gamma, prior, prior_param, M_max)
}
\description{
This function returns the log of all terms in Vpost, from m=0 up to M_max
}
