% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{log_Vprior_apprx1}
\alias{log_Vprior_apprx1}
\title{V approximation - Zeta Riemann}
\usage{
log_Vprior_apprx1(k, n_i, tol, gamma, prior, prior_param, M_max)
}
\description{
Non usare questa funzione. Non ci sono funzioni che implementino il log della zeta-Riemann. Quindi ho problemi numerici e non va
}
