% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{log_Vprior_apprx2}
\alias{log_Vprior_apprx2}
\title{V approximation - Under threshold}
\usage{
log_Vprior_apprx2(k, n_i, tol, gamma, prior, prior_param, M_max)
}
\description{
This approximation has no theoretical guarantees
}
