% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{log_Vprior_apprx3}
\alias{log_Vprior_apprx3}
\title{V approximation - Compute Upper tail}
\usage{
log_Vprior_apprx3(k, n_i, tol, gamma, prior, prior_param, M_max)
}
\description{
Non posso calcolare direttamente il log di P(M>N+K). Usala solo se gamma è molto piccolo o n_j è molto grande.
}
