% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{log_Vprior_long}
\alias{log_Vprior_long}
\title{All terms of V}
\usage{
log_Vprior_long(k, n_i, gamma, prior, prior_param, M_max)
}
\description{
This function returns the log of all terms in V, from m=0 up to M_max
}
