% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HSSM_functions.R, R/RcppExports.R
\name{lp_coverage_post}
\alias{lp_coverage_post}
\title{Log coverage prob m steps ahead}
\usage{
lp_coverage_post(m_j, n_j, r, gamma_j, prior, prior_param, absC1, absC2, M_max)

lp_coverage_post(m_j, n_j, r, gamma_j, prior, prior_param, absC1, absC2, M_max)
}
\description{
Compute the log of the coverage probability for shared species m-steps ahead
}
