% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HSSM_functions.R
\name{lp_fulljoint}
\alias{lp_fulljoint}
\title{Log full prior}
\usage{
lp_fulljoint(
  n_j,
  r,
  r_j,
  gamma_j,
  prior,
  prior_param,
  log_V,
  absC1,
  absC2,
  M_max
)
}
\description{
Compute the full joint prior probability.
}
