% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lp_fulljoint_c}
\alias{lp_fulljoint_c}
\title{Compute the logarithm of the the full joint prior probability. Cnumbers are passed as an input}
\usage{
lp_fulljoint_c(
  n_j,
  r,
  r_j,
  gamma_j,
  prior,
  prior_param,
  log_V,
  absC1,
  absC2,
  M_max
)
}
\description{
Compute the logarithm of the the full joint prior probability. Cnumbers are passed as an input
}
