% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HSSM_functions.R
\name{lp_fullpost}
\alias{lp_fullpost}
\title{Log full posterior}
\usage{
lp_fullpost(
  n_j,
  m_j,
  r,
  r_j,
  k,
  k_j,
  gamma_j,
  prior,
  prior_param,
  log_Vpost,
  absC1,
  absC2,
  M_max
)
}
\description{
Compute the full joint prior probability.
}
