% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lp_fullpost_c}
\alias{lp_fullpost_c}
\title{Compute the logarithm of the full joint posterior probability. Cnumbers are passed as an input}
\usage{
lp_fullpost_c(
  m_j,
  n_j,
  r,
  r_j,
  k,
  k_j,
  gamma_j,
  prior,
  prior_param,
  log_Vpost,
  absC1,
  absC2,
  M_max
)
}
\description{
Compute the logarithm of the full joint posterior probability. Cnumbers are passed as an input
}
