% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HSSM_functions.R
\name{p_shared_post_largen}
\alias{p_shared_post_largen}
\title{p_shared_post_largen}
\usage{
p_shared_post_largen(
  t,
  k,
  m_j,
  n_j,
  gamma,
  prior = "Poisson",
  ...,
  log_Vprior,
  Max_iter = 100
)
}
\description{
p_shared_post_largen
}
