% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{p_shared_post_largen_c}
\alias{p_shared_post_largen_c}
\title{Compute the marginal predictive posterior distribution of the number of shared species in a single point named t}
\usage{
p_shared_post_largen_c(
  t,
  k,
  m_j,
  n_j,
  gamma_j,
  prior,
  prior_param,
  log_V,
  M_max
)
}
\description{
Compute the marginal predictive posterior distribution of the number of shared species in a single point named t
}
