% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HSSM_functions.R
\name{p_shared_posterior}
\alias{p_shared_posterior}
\title{p_shared_posterior}
\usage{
p_shared_posterior(
  t,
  k,
  m_j,
  n_j,
  gamma,
  prior = "Poisson",
  ...,
  Max_iter = 100
)
}
\arguments{
\item{t}{integer, the number of shared species whose probability has to be computed.}

\item{k}{integer, the number of distinct species that have been observed in the sample of sizes given by \code{n_j}.}

\item{m_j}{an positive integer in the case of exchangeable data or a vector of size \code{d} in the case of partially exchangeable data.}

\item{n_j}{a positive scalar or a vector of positive integers that defines the size of the groups that have been alread observed. It must be of the same size of \code{m_j}.}

\item{gamma}{real valued, it must be of the same size of \code{n_j}}

\item{prior}{a string that indicates the type of prior to be used for the number of components. It can only be equal to \code{"Poisson"} or \code{"NegativeBinomial"}.}

\item{...}{the addition parameters to be used in the prior. Use \code{lambda} for the "Poisson"case (must be strictly positive) and \code{r} (positive integer) and \code{p} (real in (0,1)) for the "NegativeBinomial" case.}
}
\description{
This function computes the posterior probability that the number of shared species is equal to \code{t}.
}
