% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HSSM_functions.R
\name{p_shared_prior}
\alias{p_shared_prior}
\title{p_shared_prior}
\usage{
p_shared_prior(s, n_j, gamma, prior = "Poisson", ..., Max_iter = 100)
}
\arguments{
\item{s}{integer, the number of shared species whose probability has to be computed.}

\item{n_j}{an positive integer in the case of exchangeable data or a vector of size \code{d} in the case of partially exchangeable data.}

\item{gamma}{real valued, it must be of the same size of \code{n_j}}

\item{prior}{a string that indicates the type of prior to be used for the number of components. It can only be equal to \code{"Poisson"} or \code{"NegativeBinomial"}.}

\item{...}{the addition parameters to be used in the prior. Use \code{lambda} for the "Poisson"case (must be strictly positive) and \code{r} (positive integer) and \code{p} (real in (0,1)) for the "NegativeBinomial" case.}
}
\description{
This function computes the a priori probability that the number of shared species is equal to \code{s}.
}
