% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{PrDistinct0_c}
\alias{PrDistinct0_c}
\title{Compute the (m1,m2) step prediction probabilities of observing 0 new distinct species}
\usage{
PrDistinct0_c(n_j, m_j, r, gamma_j, prior, prior_param, absC1, absC2, M_max)
}
\description{
Compute the (m1,m2) step prediction probabilities of observing 0 new distinct species
}
