% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{PrDistinct0_d1_c}
\alias{PrDistinct0_d1_c}
\title{Compute the (m) step prediction probability of observing 0 new distinct species in a single area}
\usage{
PrDistinct0_d1_c(n_j, m_j, r_j, gamma, prior, prior_param, absCj, M_max)
}
\description{
Compute the (m) step prediction probability of observing 0 new distinct species in a single area
}
