% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{PrDistinct_1step_c}
\alias{PrDistinct_1step_c}
\title{Compute the 1 step prediction probabilities for new global distinct probabilities in a future sample of size (1,1)}
\usage{
PrDistinct_1step_c(n_j, r, r_j, rstar_j, gamma_j, prior, prior_param, M_max)
}
\description{
Compute the 1 step prediction probabilities for new global distinct probabilities in a future sample of size (1,1)
}
