% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{PrLocDistinct1_1step_c}
\alias{PrLocDistinct1_1step_c}
\title{Compute the 1 step prediction probabilities for new local distinct in the first area probabilities in a future sample of size (1,1)}
\usage{
PrLocDistinct1_1step_c(
  n_j,
  r,
  r_j,
  rstar_j,
  gamma_j,
  prior,
  prior_param,
  M_max
)
}
\description{
Compute the 1 step prediction probabilities for new local distinct in the first area probabilities in a future sample of size (1,1)
}
