% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{PrNextDistinct_d1_c}
\alias{PrNextDistinct_d1_c}
\title{Compute the 1 step prediction probability for new distinct species in a future sample of size (1)}
\usage{
PrNextDistinct_d1_c(n, r, gamma, prior, prior_param, M_max)
}
\description{
Compute the 1 step prediction probability for new distinct species in a future sample of size (1)
}
