% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{PrSh0_c}
\alias{PrSh0_c}
\title{Compute the (m1,m2) step prediction probabilities of observing 0 new shared species}
\usage{
PrSh0_c(n_j, m_j, r, r_j, gamma_j, prior, prior_param, absC1, absC2, M_max)
}
\description{
Compute the (m1,m2) step prediction probabilities of observing 0 new shared species
}
