% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{raising_factorial}
\alias{raising_factorial}
\title{Raising Factorial}
\usage{
raising_factorial(n, a)
}
\description{
\loadmathjax This function computes the rising factorial \mjseqn{(a)^{n}} using the gsl code for the Pochhammer symbol, i.e
\mjsdeqn{(a)^{n} = \frac{\Gamma(a+n)}{\Gamma(a)}}.
The raising (here denote with the upper apex) and the falling factorial (here denote with the lower apex) are related by the following relationship
\mjsdeqn{(a)_{n} = (-1)^{n}(a)^{n}}.
}
