% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HSSM_MCMC.R
\name{set_options_MCMC}
\alias{set_options_MCMC}
\title{Set options}
\usage{
set_options_MCMC(
  gamma0 = c(1, 1),
  Lambda0 = 10,
  UpdateLambda = TRUE,
  UpdateGamma = TRUE,
  L0 = 100,
  V_Lambda = 100,
  a_gamma = 1,
  b_gamma = c(1, 1),
  adapt_var_gamma_j = c(1, 1),
  adapt_var_Lambda = 1,
  dependentPrior = FALSE,
  M_max = 100,
  seed = 12345,
  gamma_guess = 1
)
}
\description{
Set options
}
