% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HSSM_MCMC.R
\name{set_options_sampler}
\alias{set_options_sampler}
\title{Set options}
\usage{
set_options_sampler(
  X_j = NULL,
  gamma0 = c(1, 1),
  beta0 = 0,
  Lambda0 = 0,
  UpdateLambda = TRUE,
  UpdateGamma = TRUE,
  UpdateBeta = TRUE,
  use_covariates = TRUE,
  L0 = 100,
  V_Lambda = 100,
  sigma2_beta = 100,
  adapt_var_gamma_j = c(1, 1),
  adapt_var_beta = 1,
  M_max = 1000,
  seed = 12345,
  gamma_guess = 1
)
}
\description{
Set options
}
