% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HSSM_functions.R
\name{Train_Test_evensize}
\alias{Train_Test_evensize}
\title{Training and Testing}
\usage{
Train_Test_evensize(data, n_train, n_test, seed)
}
\arguments{
\item{data}{M x 2 matrix with rownames indicating the species names. Unobserved species must also be included}
}
\description{
Return train and test dataset having the same group sizes
}
